+++ Testing/TinyFFR.Tests/Resources/ResourceDependencyTracker.cs
@@ -4,7 +4,8 @@
 using System.Linq;
 using Alpha = Egodystonic.TinyFFR.Resources.MockResourceAlpha;
 using Bravo = Egodystonic.TinyFFR.Resources.MockResourceBravo;
-using Impl = Egodystonic.TinyFFR.Resources.IMockResourceImplProvider;
+using AlphaImpl = Egodystonic.TinyFFR.Resources.IMockAlphaResourceImplProvider;
+using BravoImpl = Egodystonic.TinyFFR.Resources.IMockBravoResourceImplProvider;
 
 namespace Egodystonic.TinyFFR.Resources;
 
@@ -12,8 +13,8 @@ namespace Egodystonic.TinyFFR.Resources;
 class ResourceDependencyTrackerTest {
 	const int NumMockResourcesPerList = 10;
 	ResourceDependencyTracker _tracker;
-	MockResourceImplProvider _alphaImplProvider;
-	MockResourceImplProvider _bravoImplProvider;
+	MockAlphaResourceImplProvider _alphaImplProvider;
+	MockBravoResourceImplProvider _bravoImplProvider;
 	List<Alpha> _alphaResources;
 	List<Bravo> _bravoResources;
 
@@ -85,25 +86,25 @@ class ResourceDependencyTrackerTest {
 		AddDependencyAndAssert(_alphaResources[0], _bravoResources[0]);
 		AddDependencyAndAssert(_bravoResources[1], _bravoResources[0]);
 
-		Assert.AreEqual(_bravoResources[0], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0], 0));
-		Assert.AreEqual(_bravoResources[1], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0], 1));
-		Assert.AreEqual(_bravoResources[2], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0], 2));
+		Assert.AreEqual(_bravoResources[0], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0], 0));
+		Assert.AreEqual(_bravoResources[1], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0], 1));
+		Assert.AreEqual(_bravoResources[2], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0], 2));
 
-		Assert.AreEqual(_alphaResources[0], _tracker.GetNthDependentOfGivenType<Bravo, Alpha, Impl>(_bravoResources[0], 0));
-		Assert.AreEqual(_bravoResources[1], _tracker.GetNthDependentOfGivenType<Bravo, Bravo, Impl>(_bravoResources[0], 0));
-		Assert.Catch(() => _tracker.GetNthDependentOfGivenType<Bravo, Alpha, Impl>(_bravoResources[0], 1));
-		Assert.Catch(() => _tracker.GetNthDependentOfGivenType<Bravo, Bravo, Impl>(_bravoResources[0], 1));
-		Assert.Catch(() => _tracker.GetNthDependentOfGivenType<Bravo, Bravo, Impl>(_bravoResources[1], 0));
+		Assert.AreEqual(_alphaResources[0], _tracker.GetNthDependentOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[0], 0));
+		Assert.AreEqual(_bravoResources[1], _tracker.GetNthDependentOfGivenType<Bravo, Bravo, BravoImpl>(_bravoResources[0], 0));
+		Assert.Catch(() => _tracker.GetNthDependentOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[0], 1));
+		Assert.Catch(() => _tracker.GetNthDependentOfGivenType<Bravo, Bravo, BravoImpl>(_bravoResources[0], 1));
+		Assert.Catch(() => _tracker.GetNthDependentOfGivenType<Bravo, Bravo, BravoImpl>(_bravoResources[1], 0));
 
-		Assert.AreEqual(_bravoResources[0], _tracker.GetNthTargetOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0], 0));
-		Assert.Catch(() => _tracker.GetNthTargetOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0], 1));
-		Assert.Catch(() => _tracker.GetNthTargetOfGivenType<Alpha, Bravo, Impl>(_alphaResources[1], 0));
+		Assert.AreEqual(_bravoResources[0], _tracker.GetNthTargetOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0], 0));
+		Assert.Catch(() => _tracker.GetNthTargetOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0], 1));
+		Assert.Catch(() => _tracker.GetNthTargetOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[1], 0));
 		
-		Assert.AreEqual(_alphaResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Alpha, Impl>(_bravoResources[0], 0));
-		Assert.AreEqual(_alphaResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Alpha, Impl>(_bravoResources[1], 0));
-		Assert.AreEqual(_bravoResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Bravo, Impl>(_bravoResources[1], 0));
-		Assert.AreEqual(_alphaResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Alpha, Impl>(_bravoResources[2], 0));
-		Assert.AreEqual(_alphaResources[1], _tracker.GetNthTargetOfGivenType<Bravo, Alpha, Impl>(_bravoResources[2], 1));
+		Assert.AreEqual(_alphaResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[0], 0));
+		Assert.AreEqual(_alphaResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[1], 0));
+		Assert.AreEqual(_bravoResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Bravo, BravoImpl>(_bravoResources[1], 0));
+		Assert.AreEqual(_alphaResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[2], 0));
+		Assert.AreEqual(_alphaResources[1], _tracker.GetNthTargetOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[2], 1));
 
 		Assert.AreEqual(3, _tracker.GetDependents(_alphaResources[0]).Count);
 		Assert.AreEqual(1, _tracker.GetDependents(_alphaResources[1]).Count);
@@ -119,28 +120,28 @@ class ResourceDependencyTrackerTest {
 		Assert.AreEqual(2, _tracker.GetTargets(_bravoResources[1]).Count);
 		Assert.AreEqual(2, _tracker.GetTargets(_bravoResources[2]).Count);
 
-		Assert.AreEqual(3, _tracker.GetDependentsOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0]).Count);
-		Assert.AreEqual(1, _tracker.GetDependentsOfGivenType<Alpha, Bravo, Impl>(_alphaResources[1]).Count);
-		Assert.AreEqual(0, _tracker.GetDependentsOfGivenType<Alpha, Bravo, Impl>(_alphaResources[2]).Count);
+		Assert.AreEqual(3, _tracker.GetDependentsOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0]).Count);
+		Assert.AreEqual(1, _tracker.GetDependentsOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[1]).Count);
+		Assert.AreEqual(0, _tracker.GetDependentsOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[2]).Count);
 
-		Assert.AreEqual(1, _tracker.GetDependentsOfGivenType<Bravo, Alpha, Impl>(_bravoResources[0]).Count);
-		Assert.AreEqual(1, _tracker.GetDependentsOfGivenType<Bravo, Bravo, Impl>(_bravoResources[0]).Count);
-		Assert.AreEqual(0, _tracker.GetDependentsOfGivenType<Bravo, Alpha, Impl>(_bravoResources[1]).Count);
-		Assert.AreEqual(0, _tracker.GetDependentsOfGivenType<Bravo, Alpha, Impl>(_bravoResources[2]).Count);
+		Assert.AreEqual(1, _tracker.GetDependentsOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[0]).Count);
+		Assert.AreEqual(1, _tracker.GetDependentsOfGivenType<Bravo, Bravo, BravoImpl>(_bravoResources[0]).Count);
+		Assert.AreEqual(0, _tracker.GetDependentsOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[1]).Count);
+		Assert.AreEqual(0, _tracker.GetDependentsOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[2]).Count);
 
-		Assert.AreEqual(1, _tracker.GetTargetsOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0]).Count);
-		Assert.AreEqual(0, _tracker.GetTargetsOfGivenType<Alpha, Bravo, Impl>(_alphaResources[1]).Count);
-		Assert.AreEqual(0, _tracker.GetTargetsOfGivenType<Alpha, Alpha, Impl>(_alphaResources[0]).Count);
-		Assert.AreEqual(1, _tracker.GetTargetsOfGivenType<Bravo, Alpha, Impl>(_bravoResources[0]).Count);
-		Assert.AreEqual(1, _tracker.GetTargetsOfGivenType<Bravo, Alpha, Impl>(_bravoResources[1]).Count);
-		Assert.AreEqual(2, _tracker.GetTargetsOfGivenType<Bravo, Alpha, Impl>(_bravoResources[2]).Count);
+		Assert.AreEqual(1, _tracker.GetTargetsOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0]).Count);
+		Assert.AreEqual(0, _tracker.GetTargetsOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[1]).Count);
+		Assert.AreEqual(0, _tracker.GetTargetsOfGivenType<Alpha, Alpha, AlphaImpl>(_alphaResources[0]).Count);
+		Assert.AreEqual(1, _tracker.GetTargetsOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[0]).Count);
+		Assert.AreEqual(1, _tracker.GetTargetsOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[1]).Count);
+		Assert.AreEqual(2, _tracker.GetTargetsOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[2]).Count);
 
 		RemoveDependencyAndAssert(_bravoResources[1], _alphaResources[0]);
 		Assert.AreEqual(1, _tracker.GetTargets(_bravoResources[1]).Count);
 		Assert.AreEqual(2, _tracker.GetDependents(_alphaResources[0]).Count);
-		Assert.AreEqual(_bravoResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Bravo, Impl>(_bravoResources[1], 0));
-		Assert.AreEqual(_bravoResources[0], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0], 0));
-		Assert.AreEqual(_bravoResources[2], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0], 1));
+		Assert.AreEqual(_bravoResources[0], _tracker.GetNthTargetOfGivenType<Bravo, Bravo, BravoImpl>(_bravoResources[1], 0));
+		Assert.AreEqual(_bravoResources[0], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0], 0));
+		Assert.AreEqual(_bravoResources[2], _tracker.GetNthDependentOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0], 1));
 
 		Assert.Throws<ResourceDependencyException>(() => _tracker.ThrowForPrematureDisposalIfTargetHasDependents(_bravoResources[0]));
 		RemoveDependencyAndAssert(_alphaResources[0], _bravoResources[0]);
@@ -229,7 +230,7 @@ class ResourceDependencyTrackerTest {
 		AssertIteratorInvalid(iterator);
 
 		_tracker.RegisterDependency(_bravoResources[0], _alphaResources[0]);
-		var iterator2 = _tracker.GetDependentsOfGivenType<Alpha, Bravo, Impl>(_alphaResources[0]);
+		var iterator2 = _tracker.GetDependentsOfGivenType<Alpha, Bravo, BravoImpl>(_alphaResources[0]);
 		AssertIteratorValid(iterator2);
 		_tracker.DeregisterDependency(_bravoResources[0], _alphaResources[0]);
 		AssertIteratorInvalid(iterator2);
@@ -241,7 +242,7 @@ class ResourceDependencyTrackerTest {
 		AssertIteratorInvalid(iterator3);
 
 		_tracker.RegisterDependency(_bravoResources[0], _alphaResources[0]);
-		var iterator4 = _tracker.GetTargetsOfGivenType<Bravo, Alpha, Impl>(_bravoResources[0]);
+		var iterator4 = _tracker.GetTargetsOfGivenType<Bravo, Alpha, AlphaImpl>(_bravoResources[0]);
 		AssertIteratorValid(iterator4);
 		_tracker.DeregisterDependency(_bravoResources[0], _alphaResources[0]);
 		AssertIteratorInvalid(iterator4);